added type struct

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: ralee <ralee@student.42.us.org>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/27 15:30:47 by ralee             #+#    #+#             */
-/*   Updated: 2017/11/27 16:00:55 by ralee            ###   ########.fr       */
+/*   Updated: 2017/12/17 12:03:07 by ralee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(char *str)
+#include "libft.h"
+
+size_t		ft_strlen(char *str)
 {
-	int		count;
+	size_t		count;
 
 	count = 0;
 	while (str[count] != '\0')
